requirements.txt
streamlit==1.34.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
